Help channels: move validation code to __init__.py
